split some code to separate files
