+++ sanatate/scrapy_proj/helpers/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+from scrapy_proj.helpers.legal import *
+from scrapy_proj.helpers.romanian import *
+from scrapy_proj.helpers.text import *
