+++ sanatate/scrapy_proj/items/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+
+from scrapy_proj.items.act import *
+from scrapy_proj.items.contact import *
