+++ sanatate/scrapy_proj/loaders/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+
+from scrapy_proj.loaders.act import *
+from scrapy_proj.loaders.contact import *
