+++ sanatate/scrapy_proj/pipelines/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+
+from scrapy_proj.pipelines.extrameta import *
+from scrapy_proj.pipelines.post import *
